moved the visualization content in the graphviz directory
